+++ packages/contracts/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added `PermissionConditionBase` to have ERC-165 support for `IPermissionCondition` implementations.
+- Added `PermissionCondition` and `PermissionConditionUpgradeable` to have ERC-165 support for `IPermissionCondition` implementations.
 - Inherit `ProtocolVersion` and `ERC165` in `DAOFactory` and `PluginRepoFactory`.
 - Inherit `ProtocolVersion` in `DAO` and `PluginRepo`.
 - Added a `nonReentrant` modifier to the `execute` function in the `DAO` contract.